+++ scalefilter.h
@@ -82,13 +82,13 @@ public:
  * @author Daniel M. Duley (mosfet)
  */
     static QImage smoothScaleFilter(const QImage &img, const QSize &sz,
-                                    float blur=1.0,
+                                    double blur=1.0,
                                     ScaleFilterType filter=BlackmanFilter,
                                     Qt::AspectRatioMode aspectRatio =
             Qt::IgnoreAspectRatio, int page = -1, const int *currentPage = nullptr);
 
     static QImage smoothScaleFilter(const QImage &img, int dwX, int dwY,
-                                    float blur=1.0,
+                                    double blur=1.0,
                                     ScaleFilterType filter=BlackmanFilter,
                                     Qt::AspectRatioMode aspectRatio =
             Qt::IgnoreAspectRatio, int page = -1, const int *currentPage = nullptr);
